Added shielding to player
Player can now shield by holding shift. Player is slowed and this will stop all damage taken. But it is not based on direction yet. I need to change it so depending on what direction the projectile is coming, that is how damage is reduced. Also I want to make it so the player can take partial damage if they get hit by a diagonal shot while blocking in a direction that is similar (ex: player gets hit from north east but is blocking for east they will take half dmg or less). Also added placeholder sprites for the blocking player